--- a/aio/src/test-specs.ts
+++ b/aio/src/test-specs.ts
@@ -1,0 +1,4 @@
+// Find all the tests.
+const context = (<any>require).context('./', true, /\.spec\.ts$/);
+// And load the modules.
+context.keys().map(context);