--- a/aio/src/test.ts
+++ b/aio/src/test.ts
@@ -1,5 +1,6 @@
 // This file is required by karma.conf.js and loads recursively all the .spec and framework files
 
+// Test dependencies.
 import 'zone.js/dist/long-stack-trace-zone';
 import 'zone.js/dist/proxy.js';
 import 'zone.js/dist/sync-test';
@@ -12,9 +13,25 @@ import {
   platformBrowserDynamicTesting
 } from '@angular/platform-browser-dynamic/testing';
 
+// List vendors here to increase test rebuild performance.
+import '@angular/common';
+import '@angular/common/testing';
+import '@angular/core/';
+import '@angular/core/testing';
+import '@angular/platform-browser';
+import '@angular/platform-browser/testing';
+import '@angular/platform-browser/animations';
+import '@angular/platform-browser-dynamic';
+import '@angular/platform-browser-dynamic/testing';
+import '@angular/http';
+import '@angular/http/testing';
+import '@angular/animations';
+import '@angular/material';
+import '@angular/service-worker';
+import 'rxjs'; // tslint:disable-line
+
 // Unfortunately there's no typing for the `__karma__` variable. Just declare it as any.
 declare var __karma__: any;
-declare var require: any;
 
 // Prevent Karma from running prematurely.
 __karma__.loaded = function () {};
@@ -24,9 +41,9 @@ getTestBed().initTestEnvironment(
   BrowserDynamicTestingModule,
   platformBrowserDynamicTesting()
 );
+
+declare var System: any;
 // Then we find all the tests.
-const context = require.context('./', true, /\.spec\.ts$/);
-// And load the modules.
-context.keys().map(context);
-// Finally, start Karma to run the tests.
-__karma__.start();
+System.import('./test-specs.ts')
+  // Finally, start Karma to run the tests.
+  .then(() => __karma__.start());