--- a/aio/src/tsconfig.app.json
+++ b/aio/src/tsconfig.app.json
@@ -9,6 +9,7 @@
   "exclude": [
     "testing/**/*",
     "test.ts",
+    "test-specs.ts",
     "**/*.spec.ts"
   ]
 }