**Please check if the PR fulfills these requirements**
- [x] The commit message follows our guidelines: https://github.com/angular/angular/blob/master/CONTRIBUTING.md#commit
- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)


**What kind of change does this PR introduce?** (check one with "x")
```
[ ] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[ ] CI related changes
[x] Other... Please describe: performance improvement for aio unit tests
```

**What is the current behavior?** (You can also link to an open issue here)
Rebuild times for AIO unit test are somewhat slow currently: 15.809s (initial), 10.015s, 9.764s, 10.728s, 11.957s.


**What is the new behavior?**
This PR improves rebuild times: 16.551s (initial), 0.972s, 1.505s, 0.992s, 0.826s.


**Does this PR introduce a breaking change?** (check one with "x")
```
[ ] Yes
[x] No
```

If this PR contains a breaking change, please describe the impact and migration path for existing applications: ...


**Other information**:

/cc @wardbell