@intellix 2017-04-22
Also just tested this against an Angular CLI project with around ~400 tests and it's really helped the rebuilds. They're not getting stuck on 92% anymore for a couple of seconds for me